requirements.txt
asgiref==3.8.1
certifi==2024.7.4
Django==5.0.7
django-elasticsearch-dsl==8.0
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
setuptools==72.0.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
